Improved variables' declarations and made the code cleaner

+++ ft_bzero.c
@@ -6,7 +6,7 @@
 /*   By: craffate <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/11/03 18:50:22 by craffate          #+#    #+#             */
-/*   Updated: 2016/11/16 11:32:22 by craffate         ###   ########.fr       */
+/*   Updated: 2016/12/03 16:53:00 by craffate         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_bzero(void *s, size_t n)
 {
-	size_t	i;
+	unsigned int	i;
 
 	i = 0;
 	while (n-- > 0)
