+++ ft_strsub.c
@@ -6,7 +6,7 @@
 /*   By: craffate <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/11/08 14:28:33 by craffate          #+#    #+#             */
-/*   Updated: 2016/11/16 12:37:54 by craffate         ###   ########.fr       */
+/*   Updated: 2016/12/03 17:01:27 by craffate         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 char	*ft_strsub(char const *s, unsigned int start, size_t len)
 {
-	char	*s2;
-	size_t	i;
+	char			*s2;
+	unsigned int	i;
 
 	if (!s)
 		return (NULL);
